use mixin for code reuse in text_detection and seal_text_detection modules (#15109)
* use mixin for code reuse in text_detection and seal_text_detection modules

* fix input param of doc_vlm CLI

* fix
